+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Added
 - Added `domain` method to API fluent routing methods.
+- [#337](https://github.com/cloudcreativity/laravel-json-api/issues/337)
+Can now apply global scopes to JSON API resources via [adapter scopes.](./docs/basics/adapters.md#scopes)
 
 ## [1.1.0] - 2019-04-12