+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
-## Unreleased
+## [1.2.0] - 2019-06-20
 
 ### Added
 - [#360](https://github.com/cloudcreativity/laravel-json-api/issues/360)